gitignore: fix incorrect formatting
The acceptable pattern formats for .gitignore are enumerated in the
manual[1]. Surprisingly the only valid form for comments are lines
starting with a ``#`` character. A ``#`` character anywhere else will
not be parsed as one might expect.

Move the comment on ``Generated_Code`` above the pattern.

Leaving the comment where it is will a) not result in `Generated_Code``
being ignored and b) cause @rjw57's vim config to explode when opening
any file :).

[1] http://git-scm.com/docs/gitignore, "PATTERN FORMAT" section