Remove HTML comments in PDF report output.
Generated PDF reports have a large number of blank HTML comments at the start of the file before the PDF output proper starts. Removing them from the top of the file results in a PDF file which can be opened on Google Drive Viewer (am in pub so testing on phone). Since the troublesome line noted by @portableant is commented out, remove it. Closes #1004.