OAI-PMH changes
As detailed in #1178 and #1177 two issues have arisen - base url with trailing slash and invalid response of 3 elements within XML. This should fix this. I cannot fix the documentation page as I had my access rights revoked. I have not tested this on a live site, just locally. Test first on beta or training. Hope this resolves things when/if deployed CrispinF