Fixed missing region name & object certainty on CSV export
The issue was caused by a problem with an SQL query used by the SOLR
indexing function on PAS.

Region name was being set to the wrong table (region, not osRegion), and
certainty type was missing from the query.

This fix will apply to all new records, and records that are edited
after this fix. For exiting records please reindex for SOLR to be
updated with the missing data.