Fix landowner name not being removable
The form app/forms/FindSpotForm.php has two fields for landowner,
landownername and landowner, landowner being the ID of the person.

Deleting the landowner name did not remove the landowner, as such we
have added code that looks for a null or empty string in landownername
and sets landowner to null for the SQL update.