fix: editing user updates canrecord on person record
When editing canRecord on a person record, this will update the user
record as well. This is not the case the other way around.

This leads to inconsistency where the person record has canRecord, but
not the other way around.

To fix this, a new function has been created in the people class (setCanRecord)
to update the value on the people record when editing the user record.