fix - prevent RIC dropdown showing all types when NomismaID is null
When a ruler is selected that has no Nomisma ID, such as Republic,
the dropdown will be populated with all types. If one of these are
selected, this can fail the Imperial validation checks.

To prevent this, an IF statement has been added to only get
RIC dropdowns if the nomisma ID is not empty.