feat: prevent ric dropdown being filled where coin is pre-empire
The prior commit fixes this issue on the Ajax controller, which is
called when you edit a value in the form.

The CoinFormLoader is called when you load up the form, and as such this
would show all the possible RIC values in the dropdown only if republic
was already selected.