feat: refactor and fix workflow permissions
Workflow permissions are currently applied via nested if statements,
adding workflow filters to the query if not in the allowed list.

This makes it hard to edit, and in fact understand the workflow
permissions for each role. For example, "Researcher" is currently
excluded from the allowed list during this check.

As such, a new class has been added called WorkflowHandler
that will generate the filter based on a JSON array of allowed
roles for each workflow.

This change will also allow much more granularity for permissions if
wanted in the future.

Along with this, minor edits to the permissions of roles have
been made to better align with expected behaviour.