fix: can record and valid record on new user
When adding a new user via the admin area, the form settings for valid
and allowed to record do nothing.

This is as the addAction does not use the values from the checkboxes
when adding data into the var $insertData.

As such, valid and canRecord have been added to the array
initialisation for insertData and are now correctly set.