feat: fix mime detection
ImageCow will not correctly return the mime, only returning a value if
the image is a jpeg, jpg, gif or png.

This means for TIF files, no mime is returned and the function
convertTiff() is not called.

To fix this, a call to Imagick is called to get the mime for the image.