feat: ensure base image is .jpg extension
When a user uploads a valid image to the database, it will be saved
onto the database with the same extension as provided.

Pas will then resize this image into medium, small, and display folders
with a .jpg extension.

When a user clicks on the record, PAS will use the saved image name to
look in these subfolders for the image. While this works for images
uploaded as a .jpg, any other extension such as .tiff or .jpeg will
result in broken image links.

To resolve this, all images names on the database are saved with a
.jpg extension, and the original file is renamed or converted to match.