feat: add x-tiff64 to allowed mime types
x-tiff64 passes front end validation, and makes its way to the backend
as a valid tiff file.

However, Imagick has no issue with converting this file to JPEG, and
as a default option of the TIFF export on Gimp, it would be best for the
user experience to add this in.