fix: prevent search query running more than once for statistics
When a POST call is made to search on the statistics page, the user is
redirected to the same page with the dates added as URL parameters.

This means each search results in two calls to the actions in the
controller.

The controller will do the select queries on the database BEFORE seeing
if the request is a post, resulting in two groups of queries to the
database per search.

This puts undue stress on the server and database, and as such the
select queries are only ran once per search.