feat: Update date picker to latest maintained repo
The date picker used on PAS was from a closed PR for the Twitter
bootstrap framework from 2011. It has not since been maintained nor
does a repo exist for it.

While this would not be an issue if correctly working, several JS
options such as the maxDate and yearRange that are set on the date
pickers do not work. This for example allows users to enter dates far
in the future when not intended.

Of the three main bootstrap-datepicker libraries, the latest and most
maintained of these is https://github.com/eternicode/bootstrap-datepicker,
which is a fork of the original eyecon date picker found here
https://www.eyecon.ro/bootstrap-datepicker/

Moving to the new datepicker has several benefits:
    - Existing documentation
    - Working max-min dates
    - Custom formatting
    - Decade and century view
    - Parses text into correct date format (source of SQL errors
      prior)