feat: create user permission model
Currently, user permission logic is found in the individual files in
the software. This makes any permissions changes a complex, dangerous
task, which can lead to some areas of code having different
permissions.

This model will act as a unified area of permissions for the user for
future changes to the database.