feat: add request logging to bootstrap
Currently, there is no unified view of both requests and their identity
information.

While the Apache access log is useful, it does not show the username or
role of those who make a request. In addition, the access log shows all
resources loaded which adds noise to the log.

To monitor those making excessive hits such as bots and scrapers, a new
log has been added to the bootstrap class to log each request.

This will log the following:
- Username & role
- Request method
- IP address
- User agent
- Request URL