feat: limit public users to yearly search for statistics pages
As the PAS database holds over 1 million records, the statistics queries
to the database have a significant impact on the server. This can cause
the user experience to be degraded when these pages are being called
frequently.

As a hotfix to this, public users will only be able to use the yearly
links on the statistics page, rather than input their own date ranges.

This allows us to have cached data for all possible public searches,
while still allowing those who are registered to view the data with
their own dates.

Ticket: #97900