feat: allow staff to add location to map
Currently, staff who wish to be added to the contact map will need to
ask the admins to add the data to the DB.

This was not always the case; the system would attempt to get the
lat/long of the staff member from their provided address.

However, the Google Geocoder API is now a paid service, rather than
having free tier as in the past.

To rectify this, a lat & long field has been added to the form,
allowing users to add their location.

To aid in entering this data, users can select their location on a
map, which will get the lat/long of the location they click.