fix: add try/catch blocks to mailer
When email details are not set in the INI, the mail sender will throw an error.

As the mail sender is called as part of the PAS error handler, we cannot use a throw without causing a white page for the end user, and as such the PHP error_log function is used.

This change adds two try/catch blocks, one to note missing email details, and the other to catch general errors with the email sender.