feat: set file name as unique 13 digit id
When uploading a file with spaces, the name will have these trimmed by
the AJAX upload function.

However, the check for if an image already exists does not trim the
filename before checking if it already exists.

As such, these leads to cases where you can override an image if there
is already one in the directory with the same name but without spaces.

To prevent this, a unique 13-digit id is created, along with the id of
the record prepended to the file name.