feat: delete all original images
When the file is uploaded, it is converted and saved as a JPG in both
the image dir and the database.

As such, when deleting the file, PAS has no knowledge of original
non-jpg files, such as .tifs.

This change will delete all files in the user’s root image dir that has
the same filename, regardless of extension.