feat: update default format from json to search
The getFormat() function returns the format found in the params of the request, or json as a default.

This causes the SenstiveData function to filter search results as if they are json exports, not a search.

As the values from getFormat are only used in reporting, and logic, the default value has been changed to search.