fix: remove nesting array from email
The cc array holds the email details of each recipient who is copied on
the email.

Each element of this array should be an object with two properties:
email and name.

However, for the reporting user, their email details are wrapped in
another array. This prevents them from receiving the email.

This change fixes this issue by flattening the cc array and ensuring all
elements are objects.