feat: add submitter email to error report message
The error report does not contain the email address of the user who
submitted the report, insteads the user is CC'd into the email.

This caused an issue where error reports where being sent, but the
submitter was not being CC'd into the message due to an issue.

This change adds the email address into the message text, providing a
alternative means of finding who to reply to.