refactor: move map perms to UserPermissions class
The current map zoom permissions are hardcoded in the file. For
consistency, and to reduce complexity of changes, this change moves the
permission to the UsersPermissions class as ViewFullMapDetail.

For the unAuthorisedFindSpot file, force all users to not be able to
zoom in. This is as this version of the map is only shown when users
should not see any detail, and as such we can remove this permissions
check all together.