fix: null array causes email render view error
The $to array is set to null when sending a notification email for a
record with an institution of banana.

This should make the mailer use the default transaction email address.

But the $assignData value is a combination of $to and the error report
($data). Since $to is null, the array_merge function returns null as
well.

This leads to a rending error in Mailer.php later in the code,
preventing the email from being sent and showing an error to the user
because of the null array.