feat: change image filename to use oldfindid rather than id
The current method of renaming images uses the numeric ID of the record
as the prefix of the name. However, this has resulted in file names
that have little meaning for the end user.

The old ID, which is what is seen on the UI and what users are most
familiar with, includes an ID and a prefix for the institution they are
part of. By using the old ID as part of the naming scheme, we create
more usable names that are more understandable for users.