fix: Use appropriate class for thumbnail resizing
Two classes, Magick.php and MagickDust.php, share nearly identical
code, however, MagickDust includes the setUser() function, which allows
passing the username.

This function is needed when handling missing resized images,
preventing the directory for resized images from defaulting to the root
directory when the user is not logged in.

Key: PAS-91