feat: implement backend validation for images to prevent cache exhaustion
When resizing an image using ImageMagick, it consumes 8 bytes per
pixel. Consequently, when large images are uploaded, it can lead to
cache exhaustion, resulting in the following error:

cache resources exhausted ‘[…]image.jpg’ @
error/cache.c/OpenPixelCache/4083

This issue can prevent the record from opening, as the image is saved
in the slides table before the resize operation is attempted.

To address this, the current commit introduces validation to reject
images that are too large for resizing. Additionally, it returns a JSON
array in the required format for the BlueJay upload script, enabling a
front-end error message for users.